+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- ...
+- Add `testScript` option for adding flake checks based on nixosTest library.
 
 ## 0.1.0 (Jun 12, 2023)